check for empty dir

+++ Run_VEP_batch.sh
@@ -118,7 +118,7 @@ for f in $(find input/ -name '*\ *.vcf' -mmin +$waitperiod)
     mv "$f" "input/$basename"
     printf "I don't like whitespaces: $f is renamed to $basename\n"
     # Remove empty directories after files have been moved
-    if [ "${f%/*}/" != "input/" ];then
+    if [ ! "$(ls -A ${f%/*}/)" ] && [ "${f%/*}/" != "input/" ];then
       rmdir "${f%/*}/"
     fi
   done
